Tach ForecastView -> ForecastGradientView + UIViewController
